Net-Sort: Addition of ReadOnlyLists
-After testing it was realized that the way the sort functions are
operating they are modifying the input arrays. This is because since
they are reference objects they are not going to be sent as copies by
value.

So instead of sending in the array to perform the test on itself, two
local ReadOnlyLists have been created to accept the arrays that have
been initialized to ReadOnly copies. This way the same data can be
repeatedly input to the functions without worrying about changes. The
array sent in is always a copy.